@@ -11,29 +11,29 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "^20.3.15",
-    "@angular/common": "^20.3.15",
-    "@angular/compiler": "^20.3.15",
-    "@angular/core": "^20.3.15",
-    "@angular/forms": "^20.3.15",
-    "@angular/platform-browser": "^20.3.15",
-    "@angular/platform-browser-dynamic": "^20.3.15",
-    "@angular/router": "^20.3.15",
+    "@angular/animations": "^21.0.6",
+    "@angular/common": "^21.0.6",
+    "@angular/compiler": "^21.0.6",
+    "@angular/core": "^21.0.6",
+    "@angular/forms": "^21.0.6",
+    "@angular/platform-browser": "^21.0.6",
+    "@angular/platform-browser-dynamic": "^21.0.6",
+    "@angular/router": "^21.0.6",
     "rxjs": "~7.8.1",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
     "@analogjs/vite-plugin-angular": "^2.2.0",
-    "@angular-devkit/build-angular": "^20.3.13",
+    "@angular-devkit/build-angular": "^21.0.4",
     "@angular/build": "^19.2.19",
-    "@angular/cli": "^20.3.13",
-    "@angular/compiler-cli": "^20.3.15",
+    "@angular/cli": "^21.0.4",
+    "@angular/compiler-cli": "^21.0.6",
     "@testing-library/angular": "^18.1.1",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "jsdom": "^27.4.0",
-    "typescript": "~5.8.3",
+    "typescript": "~5.9.0",
     "vitest": "^4.0.16"
   }
 }
